Let PyramidPlayer know how to lose quickly
Just Drawing/Recycling over and over until you can't anymore is the
simplest and fastest way to lose quickly.  Before, when
BoardChallengeSolver couldn't find a solution, it would return the lose
quickly solution.  But the Pyramid player class should be the one that
knows how to do this and not the solver classes like
BoardChallengeSolver.  Let the solvers return no solution and let the
player handle it.