Improve game mode detection and card recognition
For Pyramid, check for goal modes before regular game modes, sometimes
it gets confused because the goal mode has a Pyramid symbol which is
what the goal mode checker looks for to see if it's a regular game.

Also one of the club images needed to be adjusted because in a test game
a 9c was misidentified as a 9s.