Refactoring - more consistent naming
Previously the code used the same terminology Microsoft Solitaire
Collection uses - the 28 cards in the pyramid are the table, and
the remaining 24 cards are initially in the deck.  But I've chosen to
call the table cards the pyramid and the deck cards the stock, so it may
avoid confusion between the 52-card deck vs the 24 cards in a pile
called the deck.