refactoring player/window classes
TriPeaks and Pyramid Solitaire have a lot of similarities so there is a
lot of overlap in the functionality between the two games for playing
and interacting with the game.  So now SolitairePlayer contains a lot of
the functionality for PyramidPlayer/TriPeaksPlayer, and MSCWindow is now
the superclass for PyramidWindow/TriPeaksWindow.