Look for solutions when there's unknown cards.
The previous behavior was to figure out what all the cards are before
searching for the solution.  Now, if the goal can be reached before
knowing all the cards (like reaching a goal score), it will stop
uncovering cards and just play the solution.

This is to avoid wasting time looking for more cards when it is
unnecessary but also to avoid having the solver become confused if it
reaches the goal score while trying to turn over face down cards.