Document non-optimal TriPeaks Score Challenges.
In TriPeaks, it's hard to guarantee the maximum possible score without
taking a long time.  Basically it's possible to reach the same overall
state but with a different order of cards removed, resulting in one
order where the score is higher but the streak is lower, and one where
the score is lower but the streak is higher.

The version with the lower score may overtake the one with the higher
score due to the bigger streak, but we don't know if that will happen
unless we examine both ways.  But doing this makes the search extremely
slow.

So right now, it always goes for the route that has the higher current
score.