Add a system report to check user settings.
I haven't been able to replicate an issue where users are reporting that
the program doesn't click on the right things. One way to start
investigating is to get details from the user.

This change prints out a System Report to the console which reports
scaling factor, DPI, screen size, and window size. I suspect things are
going wrong because the scaling factor is not 100%, which the program
explicitly does not support, but I'm not sure what the exact reason is
yet.