Make the program more language independent.
The previous version only worked in English, not Turkish. To fix this we
can look more into the positions of where to click rather than searching
images of English text to click on.

This also has the side effect of making the program work ok (for me)
when the screen resolution is at 1440x900 and the taskbar is blocking
the undo board button. Maybe this will help some users having this
issue.