add a --show-prompts command line argument
The default now lets you run the program unattended and --show-prompts
makes the program act like the previous version, where it asks you to
verify the cards and confirm playing the solution.

My reasoning is that if you really want to skip the prompts then you
probably might be frustrated typing an extra option. It's better to make
the longer command line give you the longer interaction.